requirements.txt
Flask==0.10.1
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
WTForms==2.0.2
Flask-Script
Flask-SQLAlchemy
# ----------------------
# Production Server
# ----------------------
gunicorn
Flask-Login
Flask-Restless